Move files, apply eslint and move functions into class for datalayer
