+++ index.html
@@ -13,6 +13,6 @@
 
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
+    <script type="module" src="/src/map/main.tsx"></script>
   </body>
 </html>
