+++ package.json
@@ -52,6 +52,7 @@
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.3",
+    "eslint-plugin-simple-import-sort": "^12.0.0",
     "postcss": "^8.4.32",
     "prettier": "3.1.0",
     "typescript": "^5.0.2",
