+++ src/map/components/Footer.tsx
@@ -1,8 +1,8 @@
 import { Button, Flex, Popover } from "@radix-ui/themes";
-
-import { ThemeButton } from "./Theme";
-import { useMenuStore } from "./Store";
 import { Theme } from "@radix-ui/themes";
+
+import { useMenuStore } from "./Store";
+import { ThemeButton } from "./Theme";
 
 export function Footer() {
   const theme = useMenuStore((state: any) => state.theme);
